Scramble alignment option, removed MBLD and FMC events and bug fixes
MBLD and FMC are very different events to other speed solving events,
both in the data that should be stored and handled and the interface
the user interacts with. Block Keeper cannot provide an effective
experience with its current interface. It would be better if these
events were handled by seperate apps that can provide these experiences.